move plot_convergence to plotting
