+++ gpkit/interactive/plotting.py
@@ -1,5 +1,6 @@
 """Plotting methods"""
 import matplotlib.pyplot as plt
+import numpy as np
 from ..nomials import VarKey
 from ..small_scripts import unitstr
 
@@ -226,4 +227,36 @@ def sensitivity_plot(gp, keys=None, xmax=1, yxmax=1):
     top_ax.set_xticklabels([t[1] for t in top_ticks], rotation=90)
     right_ax.set_ylim(ylim)  # this needs to occur after set_yticks().
     top_ax.set_xlim(left_ax.get_xlim())
+    return fig
+
+
+def plot_convergence(model):
+    """Plots the convergence of a signomial programming model
+
+    Arguments
+    ---------
+    model: Model
+        Signomial programming model that has already been solved
+
+    Returns
+    -------
+    matplotlib.pyplot Figure
+        Semilogy plot of variable values as functions of SP iteration #
+    """
+    newDict = {}
+
+    fig, ax = plt.subplots()
+
+    for key in model.program.gps[0].result['variables']:
+        a = np.array([])
+        for j in range(len(model.program.gps)):
+            a = np.append(a, model.program.gps[j].result['variables'][key])
+        newDict[key] = a
+        ax.semilogy(np.arange(len(newDict[key])), newDict[key], label=key.name)
+
+    box = ax.get_position()
+    ax.set_position([box.x0, box.y0, box.width * 0.8, box.height])
+    ax.set_xlabel('Number of iterations')
+    ax.set_ylabel('Normalized variable values')
+    ax.legend(loc='center left', bbox_to_anchor=(1, 0.5))
     return fig
